# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mathlers API — boxing-themed AI math competition platform
servers:
  - url: /api/v1
    description: Versioned API — v1
  - url: /api
    description: Legacy base path (backward-compat, deprecated)
...
paths:
...
  /record-cards:
    get:
      operationId: listRecordCards
      tags: [record-cards]
      summary: List all record cards
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: difficulty
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of record cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecordCard"
    post:
      operationId: createRecordCard
      tags: [record-cards]
      summary: Create a new record card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordCardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordCard"
  /record-cards/{id}:
    get:
      operationId: getRecordCard
      tags: [record-cards]
      summary: Get a record card by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Record card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordCard"
        "404":
          description: Not found
  /record-cards/scrape:
    post:
      operationId: scrapeRecordCards
      tags: [record-cards]
      summary: Trigger web scraping for new records
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeInput"
      responses:
        "200":
          description: Scrape job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeResult"
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List generated questions
      parameters:
        - in: query
          name: difficulty
          schema: { type: string }
        - in: query
          name: topic
          schema: { type: string }
        - in: query
          name: recordCardId
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /questions/generate:
    post:
      operationId: generateQuestion
      tags: [questions]
      summary: Generate a new question using AI pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuestionInput"
      responses:
        "200":
          description: Generated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}/validate:
    post:
      operationId: validateQuestion
      tags: [questions]
      summary: Re-validate a question's math correctness
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"

  # ─── STUDENTS ───────────────────────────────────────────────────
...
components:
  schemas:
...
    RecordCard:
      type: object
      required: [id, title, value, unit, category, difficulty, source, verified, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        holder: { type: ["string", "null"] }
        value: { type: number }
        unit: { type: string }
        previousValue: { type: ["number", "null"] }
        category: { type: string }
        subcategory: { type: ["string", "null"] }
        difficulty: { type: string }
        ageRange: { type: string }
        recordDate: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        source: { type: string }
        sourceUrl: { type: ["string", "null"] }
        verified: { type: boolean }
        version: { type: integer }
        createdAt: { type: string }
    RecordCardInput:
      type: object
      required: [title, value, unit, category, difficulty, source]
      properties:
        title: { type: string }
        holder: { type: string }
        value: { type: number }
        unit: { type: string }
        previousValue: { type: number }
        category: { type: string }
        subcategory: { type: string }
        difficulty: { type: string }
        ageRange: { type: string }
        recordDate: { type: string }
        location: { type: string }
        source: { type: string }
        sourceUrl: { type: string }
...
    ScrapeInput:
      type: object
      required: [sources]
      properties:
        sources:
          type: array
          items: { type: string }
        categories:
          type: array
          items: { type: string }
    ScrapeResult:
      type: object
      required: [jobId, status, message]
      properties:
        jobId: { type: string }
        status: { type: string }
        message: { type: string }
...
    Question:
      type: object
      required: [id, recordCardId, templateId, roundName, topic, difficulty, scenario, questionText, options, correctAnswer, steps, hash, validated, createdAt]
      properties:
        id: { type: integer }
        recordCardId: { type: integer }
        templateId: { type: integer }
        roundName: { type: string }
        topic: { type: string }
        difficulty: { type: string }
        scenario: { type: string }
        questionText: { type: string }
        options:
          type: array
          items: { type: string }
        correctAnswer: { type: string }
        steps:
          type: array
          items: { type: string }
        knockoutChallenge: { type: ["string", "null"] }
        hash: { type: string }
        validated: { type: boolean }
        timeLimitSeconds: { type: integer }
        pointValue: { type: integer }
        createdAt: { type: string }
        recordCard:
          $ref: "#/components/schemas/RecordCard"
    GenerateQuestionInput:
      type: object
      required: [recordCardId, difficulty, topic]
      properties:
        recordCardId: { type: integer }
        difficulty: { type: string }
        topic: { type: string }
        templateId: { type: ["integer", "null"] }
        roundName: { type: string }
        providerId: { type: ["string", "null"] }
    ValidationResult:
      type: object
      required: [valid, message]
      properties:
        valid: { type: boolean }
        message: { type: string }
        recomputedAnswer: { type: ["string", "null"] }
        steps:
          type: array
          items: { type: string }
...